Revert "Refactor authorization.js"
This reverts commit fee3e13ecaebc43ac39f0541d11998f9e89974d4.
