+++ extension/src/paymentHandler/payment-handler.js
@@ -7,7 +7,7 @@ const manualCaptureHandler = require('./manual-capture.handler')
 const cancelHandler = require('./cancel-payment.handler')
 const refundHandler = require('./refund-payment.handler')
 const pU = require('./payment-utils')
-const auth = require('../validator/authentication')
+const auth = require('../validator/authentication-validator')
 const errorMessages = require('../validator/error-messages')
 
 const { CTP_ADYEN_INTEGRATION } = require('../config/constants')
@@ -40,8 +40,7 @@ async function handlePayment(paymentObject, authToken) {
       data: paymentValidator.buildCtpErrorResponse(),
     }
 
-  const ctpProjectKey = paymentObject.custom.fields.commercetoolsProjectKey
-  if (!auth.hasValidAuthorizationHeader(ctpProjectKey, authToken)) {
+  if (!auth.isAuthorized(paymentObject, authToken)) {
     return {
       success: false,
       data: {
