+++ extension/src/utils.js
@@ -33,8 +33,14 @@ function getLogger() {
   return logger
 }
 
+function getAuthorizationRequestHeader(request) {
+  if (request.headers) return request.headers['authorization']
+  return undefined
+}
+
 module.exports = {
   collectRequestData,
   sendResponse,
   getLogger,
+  getAuthorizationRequestHeader,
 }
