+++ extension/src/validator/authentication-validator.js
@@ -1,6 +1,7 @@
 const config = require('../config/config')
 
-function hasValidAuthorizationHeader(ctpProjectKey, authTokenString) {
+function isAuthorized(paymentObject, authTokenString) {
+  const ctpProjectKey = paymentObject.custom.fields.commercetoolsProjectKey
   const isAuthEnabled = _isAuthEnabled(ctpProjectKey)
 
   if (isAuthEnabled) {
@@ -37,25 +38,6 @@ function _isAuthEnabled(ctpProjectKey) {
   return false
 }
 
-function getAuthorizationRequestHeader(request) {
-  if (request.headers) return request.headers['authorization']
-  return undefined
-}
-
-function generateBasicAuthorizationHeaderValue(ctpProjectKey) {
-  const ctpConfig = config.getCtpConfig(ctpProjectKey)
-  if (!ctpConfig && ctpConfig.username && ctpConfig.password) {
-    const username = ctpConfig.username
-    const password = ctpConfig.password
-
-    const decodeAuthToken = `${username}:${password}`
-    return `Basic ${Buffer.from(decodeAuthToken).toString('base64')}`
-  }
-  return null
-}
-
 module.exports = {
-  hasValidAuthorizationHeader,
-  getAuthorizationRequestHeader,
-  generateBasicAuthorizationHeaderValue,
+  isAuthorized,
 }
